ls and mkdir
The ls command has the -a option set in it.  It will now hide files and directories that have the hidden bit set, unless the -a option is used.

The mkdir will now also allow the user to set bits with the -m option.  These bits are what the FatFs reports with the fil info and nothing else.